Move stuff around in preperation
