+++ src/faebryk/core/cpp/include/graph/graph.hpp
@@ -70,6 +70,7 @@ class Node {
       private:
         nb::handle type;
         bool hack_cache_is_moduleinterface;
+        // std::optional<std::vector<nb::type_object>> mro{};
 
       public:
         Type(nb::handle type);
@@ -78,6 +79,9 @@ class Node {
         // TODO these are weird
         bool is_moduleinterface();
         static nb::type_object get_moduleinterface_type();
+
+        bool is_subclass(nb::type_object type);
+        bool is_subclass(std::vector<nb::type_object> types);
     };
 
   private:
@@ -121,6 +125,9 @@ class Node {
 
     void setter_no_include_parents_in_full_name(bool no_include_parents_in_full_name);
     bool getter_no_include_parents_in_full_name() const;
+
+    bool isinstance(nb::type_object type);
+    bool isinstance(std::vector<nb::type_object> types);
 
   private:
     std::unordered_set<Node_ref> get_children_all(bool include_root);
